requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
jinja2>=3.1.4
google-genai>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
pillow>=11.0.0
aiofiles>=24.1.0
httpx>=0.27.0
# LangChain / LangGraph
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.6.0
# Computer Vision (lazy import — instalado separadamente se necessario)
opencv-python-headless>=4.10.0
# Opcional — lazy import, nao bloqueia inicializacao
ultralytics>=8.3.0
# easyocr>=1.7.0
